requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data download
yfinance>=0.2.0

# Visualization
matplotlib>=3.7.0

# Progress bars
tqdm>=4.65.0

# Optional: for advanced backtesting
# talib>=0.4.0  # Requires TA-Lib C library